server/requirements.txt
absl-py==1.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.3.2
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
django-webtest==1.9.11
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dm-tree==0.1.8
etils==1.3.0
filelock==3.9.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
h5py==3.9.0
idna==3.4
importlib-resources==5.12.0
keras==2.13.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pbd==0.9
platformdirs==3.0.0
promise==2.3
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.3.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.19.0
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
